Split exec to exec and monitor
